Module google_adexchangebuyer1d3::api

Structs

Account
Configuration data for an Ad Exchange buyer account.

AccountBidderLocation
Your bidder locations that have distinct URLs.

AccountGetCall
Gets one account by ID.

AccountListCall
Retrieves the authenticated user's list of accounts.

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the AdExchangeBuyer hub.

AccountPatchCall
Updates an existing account. This method supports patch semantics.

AccountUpdateCall
Updates an existing account.

AccountsList
An account feed lists Ad Exchange buyer accounts that the user has access to. Each entry in the feed corresponds to a single buyer account.

AdExchangeBuyer
Central instance to access all AdExchangeBuyer related resource activities

BillingInfo
The configuration data for an Ad Exchange billing info.

BillingInfoGetCall
Returns the billing information for one account specified by account ID.

BillingInfoList
A billing info feed lists Billing Info the Ad Exchange buyer account has access to. Each entry in the feed corresponds to a single billing info.

BillingInfoListCall
Retrieves a list of billing information for all accounts of the authenticated user.

BillingInfoMethods
A builder providing access to all methods supported on billingInfo resources. It is not used directly, but through the AdExchangeBuyer hub.

Budget
The configuration data for Ad Exchange RTB - Budget API.

BudgetGetCall
Returns the budget information for the adgroup specified by the accountId and billingId.

BudgetMethods
A builder providing access to all methods supported on budget resources. It is not used directly, but through the AdExchangeBuyer hub.

BudgetPatchCall
Updates the budget amount for the budget of the adgroup specified by the accountId and billingId, with the budget amount in the request. This method supports patch semantics.

BudgetUpdateCall
Updates the budget amount for the budget of the adgroup specified by the accountId and billingId, with the budget amount in the request.

Creative
A creative and its classification data.

CreativeAdTechnologyProviders
There is no detailed description.

CreativeCorrections
Shows any corrections that were applied to this creative. Read-only. This field should not be set in requests.

CreativeDisapprovalReasons
The reasons for disapproval, if any. Note that not all disapproval reasons may be categorized, so it is possible for the creative to have a status of DISAPPROVED with an empty list for disapproval_reasons. In this case, please reach out to your TAM to help debug the issue. Read-only. This field should not be set in requests.

CreativeFilteringReasons
The filtering reasons for the creative. Read-only. This field should not be set in requests.

CreativeFilteringReasonsReasons
The filtering reasons.

CreativeGetCall
Gets the status for a single creative. A creative will be available 30-40 minutes after submission.

CreativeInsertCall
Submit a new creative.

CreativeListCall
Retrieves a list of the authenticated user's active creatives. A creative will be available 30-40 minutes after submission.

CreativeMethods
A builder providing access to all methods supported on creative resources. It is not used directly, but through the AdExchangeBuyer hub.

CreativeNativeAd
If nativeAd is set, HTMLSnippet and videoURL should not be set.

CreativeNativeAdAppIcon
The app icon, for app download ads.

CreativeNativeAdImage
A large image.

CreativeNativeAdLogo
A smaller image, for the advertiser logo.

CreativesList
The creatives feed lists the active creatives for the Ad Exchange buyer accounts that the user has access to. Each entry in the feed corresponds to a single creative.

DirectDeal
The configuration data for an Ad Exchange direct deal.

DirectDealGetCall
Gets one direct deal by ID.

DirectDealListCall
Retrieves the authenticated user's list of direct deals.

DirectDealMethods
A builder providing access to all methods supported on directDeal resources. It is not used directly, but through the AdExchangeBuyer hub.

DirectDealsList
A direct deals feed lists Direct Deals the Ad Exchange buyer account has access to. This includes direct deals set up for the buyer account as well as its merged stream seats.

PerformanceReport
The configuration data for an Ad Exchange performance report list.

PerformanceReportList
The configuration data for an Ad Exchange performance report list.

PerformanceReportListCall
Retrieves the authenticated user's list of performance metrics.

PerformanceReportMethods
A builder providing access to all methods supported on performanceReport resources. It is not used directly, but through the AdExchangeBuyer hub.

PretargetingConfig
There is no detailed description.

PretargetingConfigDeleteCall
Deletes an existing pretargeting config.

PretargetingConfigDimensions
Requests which allow one of these (width, height) pairs will match. All pairs must be supported ad dimensions.

PretargetingConfigExcludedPlacements
Requests containing any of these placements will not match.

PretargetingConfigGetCall
Gets a specific pretargeting configuration

PretargetingConfigInsertCall
Inserts a new pretargeting configuration.

PretargetingConfigList
There is no detailed description.

PretargetingConfigListCall
Retrieves a list of the authenticated user's pretargeting configurations.

PretargetingConfigMethods
A builder providing access to all methods supported on pretargetingConfig resources. It is not used directly, but through the AdExchangeBuyer hub.

PretargetingConfigPatchCall
Updates an existing pretargeting config. This method supports patch semantics.

PretargetingConfigPlacements
Requests containing any of these placements will match.

PretargetingConfigUpdateCall
Updates an existing pretargeting config.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.