Struct google-adexchangebuyer1d3::Account

pub struct Account {
    pub kind: Option<String>,
    pub maximum_total_qps: Option<i32>,
    pub maximum_active_creatives: Option<i32>,
    pub bidder_location: Vec<AccountBidderLocation>,
    pub cookie_matching_nid: Option<String>,
    pub number_active_creatives: Option<i32>,
    pub id: Option<i32>,
    pub cookie_matching_url: Option<String>,
}

Configuration data for an Ad Exchange buyer account.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update accounts (request|response)

get accounts (response)

list accounts (none)

patch accounts (request|response)

Fields

kind

Resource type.

maximum_total_qps

The sum of all bidderLocation.maximumQps values cannot exceed this. Please contact your technical account manager if you need to change this.

maximum_active_creatives

The maximum number of active creatives that an account can have, where a creative is active if it was inserted or bid with in the last 30 days. Please contact your technical account manager if you need to change this.

bidder_location

Your bidder locations that have distinct URLs.

cookie_matching_nid

The nid parameter value used in cookie match requests. Please contact your technical account manager if you need to change this.

number_active_creatives

The number of creatives that this account inserted or bid with in the last 30 days.

id

Account id.

cookie_matching_url

The base URL used in cookie match requests.