Struct google_adexchangebuyer1d3::AccountMethods

pub struct AccountMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on account resources. It is not used directly, but through the AdExchangeBuyer hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_adexchangebuyer1d3 as adexchangebuyer1d3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adexchangebuyer1d3::AdExchangeBuyer;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdExchangeBuyer::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.accounts();

Methods

impl<'a, C, A> AccountMethods<'a, C, A>

fn update(&self, request: Account, id: i32) -> AccountUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing account.

Arguments

request - No description provided.

id - The account id

fn patch(&self, request: Account, id: i32) -> AccountPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing account. This method supports patch semantics.

Arguments

request - No description provided.

id - The account id

fn list(&self) -> AccountListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the authenticated user's list of accounts.

fn get(&self, id: i32) -> AccountGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets one account by ID.

Arguments

id - The account id