Struct google_adexchangebuyer1d3::BillingInfoMethods

pub struct BillingInfoMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on billingInfo resources. It is not used directly, but through the AdExchangeBuyer hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_adexchangebuyer1d3 as adexchangebuyer1d3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use adexchangebuyer1d3::AdExchangeBuyer;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AdExchangeBuyer::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)` and `list(...)`
// to build up your call.
let rb = hub.billing_info();

Methods

impl<'a, C, A> BillingInfoMethods<'a, C, A>

fn get(&self, account_id: i32) -> BillingInfoGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the billing information for one account specified by account ID.

Arguments

accountId - The account id.

fn list(&self) -> BillingInfoListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of billing information for all accounts of the authenticated user.