Struct google_adexchangebuyer1d3::Creative

pub struct Creative {
    pub product_categories: Option<Vec<i32>>,
    pub status: Option<String>,
    pub agency_id: Option<String>,
    pub attribute: Option<Vec<i32>>,
    pub restricted_categories: Option<Vec<i32>>,
    pub height: Option<i32>,
    pub advertiser_name: Option<String>,
    pub html_snippet: Option<String>,
    pub buyer_creative_id: Option<String>,
    pub impression_tracking_url: Option<Vec<String>>,
    pub account_id: Option<i32>,
    pub advertiser_id: Option<Vec<String>>,
    pub kind: Option<String>,
    pub video_url: Option<String>,
    pub click_through_url: Option<Vec<String>>,
    pub corrections: Option<Vec<CreativeCorrections>>,
    pub width: Option<i32>,
    pub native_ad: Option<CreativeNativeAd>,
    pub version: Option<i32>,
    pub api_upload_timestamp: Option<String>,
    pub vendor_type: Option<Vec<i32>>,
    pub filtering_reasons: Option<CreativeFilteringReasons>,
    pub disapproval_reasons: Option<Vec<CreativeDisapprovalReasons>>,
    pub sensitive_categories: Option<Vec<i32>>,
}

A creative and its classification data.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert creatives (request|response)

list creatives (none)

get creatives (response)

Fields
product_categories: Option<Vec<i32>>

Detected product categories, if any. Read-only. This field should not be set in requests.

status: Option<String>

Creative serving status. Read-only. This field should not be set in requests.

agency_id: Option<String>

The agency id for this creative.

attribute: Option<Vec<i32>>

All attributes for the ads that may be shown from this snippet.

restricted_categories: Option<Vec<i32>>

All restricted categories for the ads that may be shown from this snippet.

height: Option<i32>

Ad height.

advertiser_name: Option<String>

The name of the company being advertised in the creative.

html_snippet: Option<String>

The HTML snippet that displays the ad when inserted in the web page. If set, videoURL should not be set.

buyer_creative_id: Option<String>

A buyer-specific id identifying the creative in this ad.

impression_tracking_url: Option<Vec<String>>

The set of urls to be called to record an impression.

account_id: Option<i32>

Account id.

advertiser_id: Option<Vec<String>>

Detected advertiser id, if any. Read-only. This field should not be set in requests.

kind: Option<String>

Resource type.

video_url: Option<String>

The url to fetch a video ad. If set, HTMLSnippet should not be set.

click_through_url: Option<Vec<String>>

The set of destination urls for the snippet.

corrections: Option<Vec<CreativeCorrections>>

Shows any corrections that were applied to this creative. Read-only. This field should not be set in requests.

width: Option<i32>

Ad width.

native_ad: Option<CreativeNativeAd>

If nativeAd is set, HTMLSnippet and videoURL should not be set.

version: Option<i32>

The version for this creative. Read-only. This field should not be set in requests.

api_upload_timestamp: Option<String>

The last upload timestamp of this creative if it was uploaded via API. Read-only. The value of this field is generated, and will be ignored for uploads. (formatted RFC 3339 timestamp).

vendor_type: Option<Vec<i32>>

All vendor types for the ads that may be shown from this snippet.

filtering_reasons: Option<CreativeFilteringReasons>

The filtering reasons for the creative. Read-only. This field should not be set in requests.

disapproval_reasons: Option<Vec<CreativeDisapprovalReasons>>

The reasons for disapproval, if any. Note that not all disapproval reasons may be categorized, so it is possible for the creative to have a status of DISAPPROVED with an empty list for disapproval_reasons. In this case, please reach out to your TAM to help debug the issue. Read-only. This field should not be set in requests.

sensitive_categories: Option<Vec<i32>>

Detected sensitive categories, if any. Read-only. This field should not be set in requests.