Struct google_adexchangebuyer1d3::CreativeCorrections

pub struct CreativeCorrections {
    pub reason: Option<String>,
    pub details: Option<Vec<String>>,
}

Shows any corrections that were applied to this creative. Read-only. This field should not be set in requests.

This type is not used in any activity, and only used as part of another schema.

Fields
reason: Option<String>

The type of correction that was applied to the creative.

details: Option<Vec<String>>

Additional details about the correction.