Struct google_adexchangebuyer1d3::CreativeNativeAd

pub struct CreativeNativeAd {
    pub body: Option<String>,
    pub advertiser: Option<String>,
    pub headline: Option<String>,
    pub image: Option<CreativeNativeAdImage>,
    pub star_rating: Option<f64>,
    pub click_tracking_url: Option<String>,
    pub call_to_action: Option<String>,
    pub app_icon: Option<CreativeNativeAdAppIcon>,
    pub logo: Option<CreativeNativeAdLogo>,
    pub impression_tracking_url: Option<Vec<String>>,
    pub price: Option<String>,
    pub store: Option<String>,
}

If nativeAd is set, HTMLSnippet and videoURL should not be set.

This type is not used in any activity, and only used as part of another schema.

Fields
body: Option<String>

A long description of the ad.

advertiser: Option<String>

no description provided

headline: Option<String>

A short title for the ad.

image: Option<CreativeNativeAdImage>

A large image.

star_rating: Option<f64>

The app rating in the app store. Must be in the range [0-5].

click_tracking_url: Option<String>

The URL to use for click tracking.

call_to_action: Option<String>

A label for the button that the user is supposed to click.

app_icon: Option<CreativeNativeAdAppIcon>

The app icon, for app download ads.

logo: Option<CreativeNativeAdLogo>

A smaller image, for the advertiser logo.

impression_tracking_url: Option<Vec<String>>

The URLs are called when the impression is rendered.

price: Option<String>

The price of the promoted app including the currency info.

store: Option<String>

The URL to the app store to purchase/download the promoted app.