Struct google_adexchangebuyer1d3::PretargetingConfig

pub struct PretargetingConfig {
    pub kind: Option<String>,
    pub verticals: Option<Vec<String>>,
    pub platforms: Option<Vec<String>>,
    pub creative_type: Option<Vec<String>>,
    pub geo_criteria_ids: Option<Vec<String>>,
    pub mobile_operating_system_versions: Option<Vec<String>>,
    pub config_id: Option<String>,
    pub excluded_user_lists: Option<Vec<String>>,
    pub is_active: Option<bool>,
    pub config_name: Option<String>,
    pub excluded_verticals: Option<Vec<String>>,
    pub dimensions: Option<Vec<PretargetingConfigDimensions>>,
    pub excluded_geo_criteria_ids: Option<Vec<String>>,
    pub supported_creative_attributes: Option<Vec<String>>,
    pub excluded_content_labels: Option<Vec<String>>,
    pub vendor_types: Option<Vec<String>>,
    pub languages: Option<Vec<String>>,
    pub user_lists: Option<Vec<String>>,
    pub mobile_devices: Option<Vec<String>>,
    pub billing_id: Option<String>,
    pub mobile_carriers: Option<Vec<String>>,
    pub excluded_placements: Option<Vec<PretargetingConfigExcludedPlacements>>,
    pub placements: Option<Vec<PretargetingConfigPlacements>>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert pretargeting config (request|response)

patch pretargeting config (request|response)

get pretargeting config (response)

update pretargeting config (request|response)

Fields

kind

The kind of the resource, i.e. "adexchangebuyer#pretargetingConfig".

verticals

Requests containing any of these vertical ids will match.

platforms

Requests matching any of these platforms will match. Possible values are PRETARGETING_PLATFORM_MOBILE, PRETARGETING_PLATFORM_DESKTOP, and PRETARGETING_PLATFORM_TABLET.

creative_type

List must contain exactly one of PRETARGETING_CREATIVE_TYPE_HTML or PRETARGETING_CREATIVE_TYPE_VIDEO.

geo_criteria_ids

Requests containing any of these geo criteria ids will match.

mobile_operating_system_versions

Requests containing any of these mobile operating system version ids will match. Values are from mobile-os.csv in the downloadable files section.

config_id

The config id; generated automatically. Leave this field blank for insert requests.

excluded_user_lists

Requests containing any of these users list ids will not match.

is_active

Whether this config is active. Required for all requests.

config_name

The name of the config. Must be unique. Required for all requests.

excluded_verticals

Requests containing any of these vertical ids will not match. Values are from the publisher-verticals.txt file in the downloadable files section.

dimensions

Requests which allow one of these (width, height) pairs will match. All pairs must be supported ad dimensions.

excluded_geo_criteria_ids

Requests containing any of these geo criteria ids will not match.

supported_creative_attributes

Creative attributes should be declared here if all creatives corresponding to this pretargeting configuration have that creative attribute. Values are from pretargetable-creative-attributes.txt in the downloadable files section.

excluded_content_labels

Requests with any of these content labels will not match. Values are from content-labels.txt in the downloadable files section.

vendor_types

Requests that allow any of these vendor ids will match. Values are from vendors.txt in the downloadable files section.

languages

Request containing any of these language codes will match.

user_lists

Requests containing any of these user list ids will match.

mobile_devices

Requests containing any of these mobile device ids will match. Values are from mobile-devices.csv in the downloadable files section.

billing_id

The id for billing purposes, provided for reference. Leave this field blank for insert requests; the id will be generated automatically.

mobile_carriers

Requests containing any of these mobile carrier ids will match. Values are from mobile-carriers.csv in the downloadable files section.

excluded_placements

Requests containing any of these placements will not match.

placements

Requests containing any of these placements will match.