Struct google_adexchangebuyer1d3::DirectDeal

pub struct DirectDeal {
    pub advertiser: Option<String>,
    pub kind: Option<String>,
    pub buyer_account_id: Option<String>,
    pub account_id: Option<i32>,
    pub name: Option<String>,
    pub currency_code: Option<String>,
    pub allows_alcohol: Option<bool>,
    pub deal_tier: Option<String>,
    pub fixed_cpm: Option<String>,
    pub start_time: Option<String>,
    pub publisher_blocks_overriden: Option<bool>,
    pub end_time: Option<String>,
    pub seller_network: Option<String>,
    pub id: Option<String>,
    pub private_exchange_min_cpm: Option<String>,
}

The configuration data for an Ad Exchange direct deal.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get direct deals (response)

list direct deals (none)

Fields
advertiser: Option<String>

The name of the advertiser this deal is for.

kind: Option<String>

Resource type.

buyer_account_id: Option<String>

The account id that this deal was negotiated for. It is either the buyer or the client that this deal was negotiated on behalf of.

account_id: Option<i32>

The account id of the buyer this deal is for.

name: Option<String>

Deal name.

currency_code: Option<String>

The currency code that applies to the fixed_cpm value. If not set then assumed to be USD.

allows_alcohol: Option<bool>

Whether the publisher for this deal is eligible for alcohol ads.

deal_tier: Option<String>

The deal type such as programmatic reservation or fixed price and so on.

fixed_cpm: Option<String>

The fixed price for this direct deal. In cpm micros of currency according to currency_code. If set, then this deal is eligible for the fixed price tier of buying (highest priority, pay exactly the configured fixed price).

start_time: Option<String>

Start time for when this deal becomes active. If not set then this deal is active immediately upon creation. In seconds since the epoch.

publisher_blocks_overriden: Option<bool>

If true, the publisher has opted to have their blocks ignored when a creative is bid with for this deal.

end_time: Option<String>

End time for when this deal stops being active. If not set then this deal is valid until manually disabled by the publisher. In seconds since the epoch.

seller_network: Option<String>

The name of the publisher offering this direct deal.

id: Option<String>

Deal id.

private_exchange_min_cpm: Option<String>

The minimum price for this direct deal. In cpm micros of currency according to currency_code. If set, then this deal is eligible for the private exchange tier of buying (below fixed price priority, run as a second price auction).